+++ src/afterlife/config.py
@@ -8,6 +8,8 @@ class Config:
     unrotated_key_days: int = 180
     oauth_stale_days: int = 90
     inactive_admin_days: int = 30
+    privilege_drift_days: int = 90
+    privilege_drift_min_unused: int = 5
 
 
 DEFAULT = Config()